backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-dateutil 